+++ seed/ts-sdk/exhaustive/dev-dependencies/src/api/resources/types/resources/object/types/ObjectWithOptionalField.ts
@@ -15,4 +15,5 @@ export interface ObjectWithOptionalField {
     list?: string[];
     set?: Set<string>;
     map?: Record<number, string>;
+    bigint?: string;
 }